MA-19152 [#imx-2507] Align tile status buffer to 256 Bytes
Since the resolve engine require the alignment for tile status buffer,
so need to align the size to 256 Bytes. This patch has two changes.
One is to align to 256 bytes, the other is to add the extra 256 bytes.
Adding the extra 256 bytes by default is for new HW feature.
But the first change (align to 256 bytes) is needed for current HW.

Signed-off-by: Richard Liu <xuegang.liu@nxp.com>